tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/145.html

<!DOCTYPE html>
<html><head>
  <title>scheduler: SVG inline script adding text to empty script </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test();
</script>
<svg>
<script></script>
</svg>
<script>
t.step(function() {
  var s = document.querySelector("svg > script");
  s.textContent = "t.step(function() {log('inline script #1')})";
});
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["inline script #1"]);
  t.done();
});
</script>
